Async now works perfectly.
To make it work I had to set up a "partial buffer" data structure which
collects up the last line of the last stdout chunk read and buffering
it. Then it will combine it with the beginning of the next upcoming
chunk read.

This is to make sure that each line in the output is a complete line and
there are no broken lines in the output due to incomplete lines being
read.

Overall I'm super happy with how fast and snappy the workflow is now.