Add `config.lua` for configuration options
Any configuration options have defaults and can be overridden through
the user's `opts` table invoked by the plugin manager -> `init.setup()`
including keymaps.

Note: For now the only keymap configuration is for the sendmail command
but I will extend it going forward so that it is as modular and
configurable as possible.